Added list functionality and label on hover code
List is using controls, probably better to just have it as separate
elements that link to the markers. Would like them to be drag, drop and
lock-able. The hover was included to avoid popups, but can't add
separate click and double click events on the markers. May be useful for
later iteration though...